+++ CHANGELOG.md
@@ -646,6 +646,14 @@ Simple chat notation
 
 ## In pre-release
 
+### `0.81.0` _(2025-01-!!!!!!)_
+
+Editing and templates
+
+-   `fetch` is passed through `ExecutionTools` to allow proxying in browser
+-   Make new package [`@promptbook/editable`](https://www.npmjs.com/package/@promptbook/editable) and move misc editing tools there
+-   Make new package [`@promptbook/templates`](https://www.npmjs.com/package/@promptbook/templates) and add function `getBookTemplate`
+
 ## Drafts
 
 ### `0..0` _(2024--)_